requirements.txt
langchain==0.1.8
langchain-community==0.0.21
PyPDF2==3.0.1
streamlit==1.31.1
faiss-cpu
openai
huggingface_hub
sentence-transformers==2.2.2
InstructorEmbedding==1.0.1
python-dotenv